+++ src/Lykke.Service.Assets.Repositories/Entities/AssetConditionLayerEntity.cs
@@ -10,12 +10,12 @@ namespace Lykke.Service.Assets.Repositories.Entities
 
         public AssetConditionLayerEntity(string partitionKey, string rowKey, decimal priority, string description) : base(partitionKey, rowKey)
         {
-            Priority = priority;
+            Priority = (double)priority;
             Description = description;
         }
 
         public string Id => RowKey;
-        public decimal Priority { get; set; }
+        public double Priority { get; set; }
         public string Description { get; set; }
     }
 }
